feat: builder plugin swc base
