+++ packages/builder/plugin-swc-base/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  extends: ['@modern-js'],
+  ignorePatterns: ['compiled/', 'fixtures/**', 'tests/**', 'vitest.config.ts'],
+  parserOptions: {
+    project: require.resolve('./tsconfig.json'),
+  },
+};
