+++ packages/builder/plugin-swc-base/CHANGELOG.md
@@ -1,0 +1,5 @@
+# @modern-js/builder-plugin-swc-base
+
+### Major Changes
+
+Initial Release
