+++ packages/builder/plugin-swc-base/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@modern-js/builder-plugin-swc",
-  "version": "2.10.0",
+  "name": "@modern-js/builder-plugin-swc-base",
+  "version": "2.9.0",
   "description": "SWC plugin for builder in Modern.js",
   "main": "./dist/index.js",
   "types": "./src/index.ts",
@@ -13,32 +13,7 @@
     "prepublishOnly": "only-allow-pnpm"
   },
   "exports": {
-    ".": "./dist/index.js",
-    "./loader": {
-      "types": "./dist/loader.d.ts",
-      "default": "./dist/loader.js"
-    },
-    "./plugin": {
-      "types": "./dist/plugin.d.ts",
-      "default": "./dist/plugin.js"
-    },
-    "./binding": {
-      "types": "./dist/binding.d.ts",
-      "default": "./dist/binding.js"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      "loader": [
-        "./dist/loader.d.ts"
-      ],
-      "plugin": [
-        "./dist/plugin.d.ts"
-      ],
-      "binding": [
-        "./dist/binding.d.ts"
-      ]
-    }
+    ".": "./dist/index.js"
   },
   "engines": {
     "node": ">=12.0.0"
@@ -50,7 +25,6 @@
     "@modern-js/builder-webpack-provider": "workspace:*",
     "@modern-js/utils": "workspace:*",
     "@scripts/vitest-config": "workspace:*",
-    "@swc/core": "1.3.42",
     "@types/babel__core": "^7.1.20",
     "magic-string": "^0.26.2",
     "source-map": "^0.7.4",
@@ -58,10 +32,14 @@
     "webpack": "^5.76.2"
   },
   "dependencies": {
-    "@modern-js/builder-plugin-swc-base": "workspace:*",
+    "@babel/core": "^7.20.5",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.17.12",
+    "@babel/preset-typescript": "^7.17.12",
     "@modern-js/builder-shared": "workspace:*",
-    "@modern-js/swc-plugins": "0.2.0",
-    "@modern-js/utils": "workspace:*"
+    "@modern-js/utils": "workspace:*",
+    "@swc/helpers": "0.4.12",
+    "core-js": "^3.26.0"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/",
