+++ packages/builder/plugin-swc-base/src/index.ts
@@ -1,0 +1,4 @@
+export * from './apply';
+export * from './checkPolyfillPlugin';
+export * from './createLoader';
+export * from './types';
