+++ packages/builder/plugin-swc/tests/loader.test.ts
@@ -1,6 +1,7 @@
-import { describe, it } from 'vitest';
+import { describe } from 'vitest';
 import { LoaderContext, LoaderDefinitionFunction } from 'webpack';
-import { createLoader } from '../src/loader';
+import { createLoader } from '@modern-js/builder-plugin-swc-base';
+import { Compiler } from '../src/binding';
 
 const mockSwcLoaderRunner = (): [
   Promise<ReturnType<LoaderDefinitionFunction>>,
@@ -25,7 +26,7 @@ const mockSwcLoaderRunner = (): [
 };
 
 describe('should handle loader interface correctly', async done => {
-  const loader = createLoader();
+  const loader = createLoader(Compiler);
 
   const [finish, runner] = mockSwcLoaderRunner();
