+++ tests/e2e/builder/cases/swc/CHANGELOG.md
@@ -1,0 +1,5 @@
+# @e2e/webpack-builder-swc
+
+## 2.9.0
+
+## 2.8.0
