+++ tests/e2e/builder/cases/swc/package.json
@@ -1,0 +1,22 @@
+{
+  "private": true,
+  "name": "@e2e/webpack-builder-swc",
+  "version": "2.9.0",
+  "scripts": {
+    "test": "playwright test"
+  },
+  "dependencies": {
+    "react": "^18.0.2",
+    "react-dom": "^18.0.2"
+  },
+  "devDependencies": {
+    "@modern-js/e2e": "workspace:*",
+    "@modern-js/builder": "workspace:*",
+    "@modern-js/builder-webpack-provider": "workspace:*",
+    "@modern-js/builder-plugin-swc": "workspace:*",
+    "@types/node": "^14",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "typescript": "^4"
+  }
+}
