+++ tests/e2e/builder/cases/swc/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "@modern-js/tsconfig/base",
+  "compilerOptions": {
+    "declaration": false,
+    "jsx": "react-jsx",
+    "baseUrl": "./",
+    "outDir": "./dist",
+    "paths": {
+      "@/*": ["./src/*"],
+      "@shared/*": ["./shared/*"],
+      "@scripts/*": ["../../scripts/*"]
+    }
+  },
+  "include": ["src", "*.test.ts", "cssModules.test.tss"]
+}
